Refactored vuex to use modules
