+++ src/store/types.ts
@@ -1,0 +1,3 @@
+export interface RootState {
+  showTutorial: boolean;
+}
